cluster/ai-ops-agent/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP Client
httpx==0.25.2

# Vector Database Client
qdrant-client==1.7.0

# YAML Processing
pyyaml==6.0.1

# Python Standard Library Backports
python-multipart==0.0.6

# Optional: PostgreSQL (for production ML logging)
# asyncpg==0.29.0

# Optional: Prometheus metrics
# prometheus-client==0.19.0
